Updated connect template directory structure
