+++ scaffolds/nextjs-magic-connect/template/.env.example
@@ -1,5 +1,4 @@
-MAGIC_API_KEY="pk_live_8D6C562ABCA3140A"
 NEXT_PUBLIC_OPTIMISM_RPC_URL=https://goerli.optimism.io/
 NEXT_PUBLIC_POLYGON_RPC_URL=https://rpc-mumbai.maticvigil.com/
 NEXT_PUBLIC_ETHEREUM_RPC_URL=https://eth-goerli.g.alchemy.com/v2/3jKhhva6zBqwp_dnwPlF4d0rFZhu2pjD/
-NEXT_PUBLIC_MAGIC_API_KEY=pk_live_8D6C562ABCA3140A
+NEXT_PUBLIC_MAGIC_API_KEY=
