+++ scaffolds/nextjs-magic-connect/template/.eslintrc.json
@@ -1,0 +1,3 @@
+{
+  "extends": "next/core-web-vitals"
+}
