+++ scaffolds/nextjs-magic-connect/template/package.json
@@ -1,24 +1,26 @@
 {
-  "name": "<%= projectName %>",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "@types/node": "18.15.11",
-    "@types/react": "18.0.35",
-    "@types/react-dom": "18.0.11",
-    "eslint": "8.38.0",
-    "eslint-config-next": "13.3.0",
-    "magic-sdk": "^16.2.0",
-    "next": "13.3.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "typescript": "5.0.4",
-    "web3": "^1.9.0"
-  }
+	"name": "magic-nextjs",
+	"version": "0.1.0",
+	"private": true,
+	"scripts": {
+		"dev": "next dev",
+		"build": "next build",
+		"start": "next start",
+		"lint": "next lint --fix"
+	},
+	"dependencies": {
+		"@types/node": "18.15.11",
+		"@types/react": "18.0.35",
+		"@types/react-dom": "18.0.11",
+		"eslint-config-next": "13.3.0",
+		"magic-sdk": "^16.2.0",
+		"next": "13.3.0",
+		"react": "18.2.0",
+		"react-dom": "18.2.0",
+		"typescript": "5.0.4",
+		"web3": "^1.9.0"
+	},
+	"devDependencies": {
+		"eslint": "^8.45.0"
+	}
 }
