+++ scaffolds/nextjs-magic-connect/template/src/components/ui/FormButton.tsx
@@ -1,0 +1,17 @@
+import React from 'react'
+
+interface Props {
+	children: React.ReactNode
+	onClick: () => void
+	disabled: boolean
+}
+
+const FormButton = ({children, onClick, disabled}: Props) => {
+	return (
+		<button className='form-button' disabled={disabled} onClick={onClick}>
+			{children}
+		</button>
+	)
+}
+
+export default FormButton
