--- scaffolds/nextjs-magic-connect/template/src/components/ui/card-header.tsx
@@ -1,16 +1,0 @@
-import React from 'react';
-
-interface Props {
-  children: React.ReactNode;
-  id: string;
-}
-
-const CardHeader = ({ children, id }: Props) => {
-  return (
-    <h1 className="card-header" id={id}>
-      {children}
-    </h1>
-  );
-};
-
-export default CardHeader;
