+++ scaffolds/nextjs-magic-connect/template/src/components/ui/spacer.tsx
@@ -1,7 +1,7 @@
-import React from "react"
+import React from 'react'
 
-const Spacer = ({ size }: { size: number }) => {
-  return <div style={{ height: `${size}px` }} />
+const Spacer = ({size}: {size: number}) => {
+	return <div style={{height: `${size}px`}} />
 }
 
 export default Spacer
