+++ scaffolds/nextjs-magic-connect/template/src/pages/_app.tsx
@@ -1,11 +1,11 @@
-import "@/styles/globals.css"
-import type { AppProps } from "next/app"
-import MagicProvider from "@/context/magic-context"
+import '@/styles/globals.css'
+import type {AppProps} from 'next/app'
+import MagicProvider from '@/components/magic/MagicProvider'
 
-export default function App({ Component, pageProps }: AppProps) {
-  return (
-    <MagicProvider>
-      <Component {...pageProps} />
-    </MagicProvider>
-  )
+export default function App({Component, pageProps}: AppProps) {
+	return (
+		<MagicProvider>
+			<Component {...pageProps} />
+		</MagicProvider>
+	)
 }
